Changed kinds of reals throughout to be explicitly double precision
- Kind specified in detypes.f90
- All incoming real values need to be declared as kind=dp
- Should play nicely with MPI data types...

+++ Fortran/source/mutation.f90
@@ -7,22 +7,22 @@ implicit none
 private
 public mutate, init_FjDE
 
-real, parameter :: tau=0.1, Fl=0.1, Fu=0.9 !jDE control parameters as in Brest et al 2006
+real(dp), parameter :: tau=0.1, Fl=0.1, Fu=0.9 !jDE control parameters as in Brest et al 2006
 
 contains
 
   subroutine mutate(X, V, n, run_params, trialF) 
     type(population), intent(in) :: X                      !current generation of target vectors
     type(codeparams), intent(in) :: run_params
-    real, dimension(run_params%D), intent(out) :: V        !donor vector
+    real(dp), dimension(run_params%D), intent(out) :: V        !donor vector
     integer, intent(in) :: n                               !index of current vector
-    real, intent(out) :: trialF
+    real(dp), intent(out) :: trialF
 
     if(run_params%DE%jDE) then
        trialF = newF(X, n)
        V = jDEmutation(X, n, run_params, trialF)
     else
-       trialF = 0.
+       trialF = 0.0_dp
        V = genmutation(X, n, run_params)
     endif
 
@@ -34,22 +34,22 @@ contains
     type(population), intent(in) :: X                !current generation of target vectors
     integer, intent(in) :: n                         !index of current vector
     type(codeparams), intent(in) :: run_params
-    real, dimension(run_params%D) :: genmutation     !donor vector
+    real(dp), dimension(run_params%D) :: genmutation     !donor vector
     integer, dimension(2*run_params%DE%Fsize) :: r   !index of random vectors X_J, X_K
     integer ri                                       !index of (random or current) vector X_I
     integer,  dimension(1) :: rbest                  !index of best vector (array to make minloc() happy)
     integer q                                        !use to iterate over the scale factors F(q)
-    real, dimension(run_params%D) :: sumF            !the summed difference vector over the F's
+    real(dp), dimension(run_params%D) :: sumF        !the summed difference vector over the F's
 
     !assign rbest
-    if (run_params%DE%lambda .gt. 0.0) then
+    if (run_params%DE%lambda .gt. 0.0_dp) then
        rbest = minloc(X%values) 
     else
        rbest = (/n/)                          !don't want to restrict options for ri, r(q)
     end if
 
     !assign ri
-    if(run_params%DE%current .or. (run_params%DE%lambda .eq. 1.0)) then 
+    if(run_params%DE%current .or. (run_params%DE%lambda .eq. 1.0_dp)) then 
        ri = n                                !use current target vector for mutation (for lambda=1, don't need unique ri)
     else                                     !for rand/ or rand-to-best/
        do                                    !choose random ri not equal to n or rbest (if lambda>0)
@@ -75,7 +75,7 @@ contains
     end do
 
     !find the difference vector associated with the F's:
-    sumF(:) = 0.0
+    sumF(:) = 0.0_dp
     do q=1, run_params%DE%Fsize
        sumF(:) = run_params%DE%F(q)*(X%vectors(r(2*q-1),:) - X%vectors(r(2*q),:)) + sumF(:)
     end do
@@ -91,8 +91,8 @@ contains
     type(population), intent(in) :: X
     integer, intent(in) :: n 
     type(codeparams), intent(in) :: run_params
-    real, intent(in) :: trialF
-    real, dimension(run_params%D) :: jDEmutation
+    real(dp), intent(in) :: trialF
+    real(dp), dimension(run_params%D) :: jDEmutation
     integer :: r1, r2, r3
 
     !set each D-dimensional donor vector in V by picking 3 separate random vectors from X
@@ -117,8 +117,8 @@ contains
   function newF(X, n)              !choose a trial value for the F parameter
     type(population), intent(in) :: X
     integer, intent(in) :: n
-    real :: newF
-    real :: rand1, rand2
+    real(dp) :: newF
+    real(dp) :: rand1, rand2
 
     call random_number(rand1)
     if (rand1 .lt. tau) then     
@@ -132,8 +132,8 @@ contains
 
   function init_FjDE(run_params)
     type(codeparams), intent(in) ::run_params
-    real, dimension(run_params%mpipopchunk) :: init_FjDE
-    real, dimension(run_params%mpipopchunk) :: rand
+    real(dp), dimension(run_params%mpipopchunk) :: init_FjDE
+    real(dp), dimension(run_params%mpipopchunk) :: rand
     
     call random_number(rand)
     init_FjDE =  Fl + rand*Fu
@@ -144,8 +144,8 @@ contains
   subroutine random_int(harvest, min, max) !choose a random integer between min and max, inclusive
     integer, intent(out) :: harvest
     integer, intent(in) :: min, max
-    real :: range
-    real :: r
+    real(dp) :: range
+    real(dp) :: r
    
     range = max - min
     call random_number(r)
